+++ verik-plugin/src/main/kotlin/io/verik/plugin/main/VerikPlugin.kt
@@ -8,6 +8,8 @@ import io.verik.compiler.main.VerikCompilerException
 import io.verik.compiler.main.VerikCompilerMain
 import io.verik.importer.main.VerikImporterException
 import io.verik.importer.main.VerikImporterMain
+import io.verik.plugin.`object`.VerikDomainObject
+import io.verik.plugin.`object`.VerikDomainObjectImpl
 import org.gradle.api.Action
 import org.gradle.api.GradleException
 import org.gradle.api.JavaVersion
@@ -17,19 +19,21 @@ import org.gradle.api.Task
 import org.gradle.api.plugins.JavaPluginExtension
 
 /**
- * Top level plugin class that registers the verik and verikImport gradle tasks for the compiler and importer
- * respectively.
+ * Top level plugin class that registers the gradle tasks.
  */
 @Suppress("unused")
 class VerikPlugin : Plugin<Project> {
 
     override fun apply(project: Project) {
         project.dependencies.add("implementation", "io.verik:verik-core:${ConfigUtil.getVersion()}")
-        val verikBuildTask = createVerikBuildTask(project)
-        val verikCompileTask = createVerikCompileTask(project)
-        val verikImportTask = createVerikImportTask(project)
-        verikBuildTask.dependsOn(verikCompileTask)
+        val extension = VerikDomainObjectImpl()
+        project.extensions.add(VerikDomainObject::class.java, "verik", extension)
+
+        val verikImportTask = createVerikImportTask(project, extension)
+        val verikCompileTask = createVerikCompileTask(project, extension)
+        val verikTask = createVerikTask(project, extension)
         verikCompileTask.dependsOn(verikImportTask)
+        verikTask.dependsOn(verikCompileTask)
 
         val javaPluginExtension = project.extensions.getByType(JavaPluginExtension::class.java)
         javaPluginExtension.sourceCompatibility = JavaVersion.VERSION_1_8
@@ -39,77 +43,79 @@ class VerikPlugin : Plugin<Project> {
         }
     }
 
-    private fun createVerikBuildTask(project: Project): Task {
-        val targetContainer = TargetContainer(
-            project.objects.polymorphicDomainObjectContainer(TargetDomainObject::class.java)
-        )
-        project.extensions.add("verikBuild", targetContainer)
-        val task = project.tasks.create("verikBuild") {
-            it.doLast(VerikBuildAction(project, targetContainer))
-        }
-
-        task.group = "verik"
-        return task
-    }
-
     @Suppress("DuplicatedCode")
-    private fun createVerikCompileTask(project: Project): Task {
-        val extension = project.extensions.create("verikCompile", VerikCompilerPluginExtension::class.java)
-        val task = project.tasks.create("verikCompile") {
-            it.doLast(VerikCompileAction(project, extension))
-        }
-
-        task.group = "verik"
-        task.inputs.property("toolchain", { ConfigUtil.getToolchain() })
-        task.inputs.property("timescale", { extension.timescale })
-        task.inputs.property("entryPoints", { extension.entryPoints })
-        task.inputs.property("enableDeadCodeElimination", { extension.enableDeadCodeElimination })
-        task.inputs.property("labelLines", { extension.labelLines })
-        task.inputs.property("indentLength", { extension.indentLength })
-        task.inputs.property("wrapLength", { extension.wrapLength })
-        task.inputs.property("suppressedWarnings", { extension.suppressedWarnings })
-        task.inputs.property("promotedWarnings", { extension.promotedWarnings })
-        task.inputs.property("maxErrorCount", { extension.maxErrorCount })
-        task.inputs.property("debug", { extension.debug })
-        task.inputs.files({ VerikCompilerConfigBuilder.getSourceSetConfigs(project).flatMap { it.files } })
-        task.outputs.dir(VerikCompilerConfigBuilder.getBuildDir(project))
-        return task
-    }
-
-    @Suppress("DuplicatedCode")
-    private fun createVerikImportTask(project: Project): Task {
-        val extension = project.extensions.create("verikImport", VerikImporterPluginExtension::class.java)
+    private fun createVerikImportTask(project: Project, extension: VerikDomainObjectImpl): Task {
         val task = project.tasks.create("verikImport") {
             it.doLast(VerikImportAction(project, extension))
         }
         task.group = "verik"
         task.inputs.property("toolchain", { ConfigUtil.getToolchain() })
-        task.inputs.property("includeDirs", { extension.includeDirs.joinToString() })
-        task.inputs.property("enablePreprocessorOutput", { extension.enablePreprocessorOutput })
-        task.inputs.property("suppressedWarnings", { extension.suppressedWarnings })
-        task.inputs.property("promotedWarnings", { extension.promotedWarnings })
         task.inputs.property("maxErrorCount", { extension.maxErrorCount })
         task.inputs.property("debug", { extension.debug })
-        task.inputs.files({ extension.importedFiles })
+        task.inputs.property("includeDirs", { extension.import.includeDirs.joinToString() })
+        task.inputs.property("enablePreprocessorOutput", { extension.import.enablePreprocessorOutput })
+        task.inputs.property("suppressedWarnings", { extension.import.suppressedWarnings })
+        task.inputs.property("promotedWarnings", { extension.import.promotedWarnings })
+
+        task.inputs.files({ extension.import.importedFiles })
         task.outputs.dirs({
-            if (extension.importedFiles.isNotEmpty()) {
+            if (extension.import.importedFiles.isNotEmpty()) {
                 VerikImporterConfigBuilder.getBuildDir(project)
             } else listOf()
         })
         return task
     }
 
-    private class VerikBuildAction(
+    @Suppress("DuplicatedCode")
+    private fun createVerikCompileTask(project: Project, extension: VerikDomainObjectImpl): Task {
+        val task = project.tasks.create("verikCompile") {
+            it.doLast(VerikCompileAction(project, extension))
+        }
+
+        task.group = "verik"
+        task.inputs.property("toolchain", { ConfigUtil.getToolchain() })
+        task.inputs.property("maxErrorCount", { extension.maxErrorCount })
+        task.inputs.property("debug", { extension.debug })
+        task.inputs.property("timescale", { extension.compile.timescale })
+        task.inputs.property("entryPoints", { extension.compile.entryPoints })
+        task.inputs.property("enableDeadCodeElimination", { extension.compile.enableDeadCodeElimination })
+        task.inputs.property("labelLines", { extension.compile.labelLines })
+        task.inputs.property("indentLength", { extension.compile.indentLength })
+        task.inputs.property("wrapLength", { extension.compile.wrapLength })
+        task.inputs.property("suppressedWarnings", { extension.compile.suppressedWarnings })
+        task.inputs.property("promotedWarnings", { extension.compile.promotedWarnings })
+
+        task.inputs.files({ VerikCompilerConfigBuilder.getSourceSetConfigs(project).flatMap { it.files } })
+        task.outputs.dir(VerikCompilerConfigBuilder.getBuildDir(project))
+        return task
+    }
+
+    private fun createVerikTask(project: Project, extension: VerikDomainObjectImpl): Task {
+        val task = project.tasks.create("verik") {
+            it.doLast(VerikAction(project, extension))
+        }
+
+        task.group = "verik"
+        return task
+    }
+
+    private class VerikImportAction(
         private val project: Project,
-        private val targetContainer: TargetContainer
+        private val extension: VerikDomainObjectImpl
     ) : Action<Task> {
 
-        override fun execute(task: Task) {}
+        override fun execute(task: Task) {
+            try {
+                VerikImporterMain.run(VerikImporterConfigBuilder.getConfig(project, extension))
+            } catch (exception: VerikImporterException) {
+                throw GradleException("Verik import failed")
+            }
+        }
     }
 
     private class VerikCompileAction(
         private val project: Project,
-        private val extension: VerikCompilerPluginExtension
+        private val extension: VerikDomainObjectImpl
     ) : Action<Task> {
 
         override fun execute(task: Task) {
@@ -121,17 +127,11 @@ class VerikPlugin : Plugin<Project> {
         }
     }
 
-    private class VerikImportAction(
+    private class VerikAction(
         private val project: Project,
-        private val extension: VerikImporterPluginExtension
+        private val extension: VerikDomainObjectImpl
     ) : Action<Task> {
 
-        override fun execute(task: Task) {
-            try {
-                VerikImporterMain.run(VerikImporterConfigBuilder.getConfig(project, extension))
-            } catch (exception: VerikImporterException) {
-                throw GradleException("Verik import failed")
-            }
-        }
+        override fun execute(task: Task) {}
     }
 }