+++ verik-plugin/src/main/kotlin/io/verik/plugin/object/VerikCompileDomainObject.kt
@@ -1,0 +1,20 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.verik.plugin.`object`
+
+/**
+ * Interface for domain object that configures the compiler.
+ */
+interface VerikCompileDomainObject {
+
+    var timescale: String
+    var entryPoints: List<String>
+    var enableDeadCodeElimination: Boolean
+    var labelLines: Boolean
+    var indentLength: Int
+    var wrapLength: Int
+    var suppressedWarnings: List<String>
+    var promotedWarnings: List<String>
+}